Fix win warning

+++ src/include/common/types/types.h
@@ -33,8 +33,8 @@ constexpr vector_idx_t INVALID_VECTOR_IDX = UINT32_MAX;
 using block_idx_t = uint64_t;
 using field_idx_t = uint64_t;
 using struct_entry_t = int64_t;
-using struct_field_idx_t = uint64_t;
-constexpr struct_field_idx_t INVALID_STRUCT_FIELD_IDX = UINT64_MAX;
+using struct_field_idx_t = uint32_t;
+constexpr struct_field_idx_t INVALID_STRUCT_FIELD_IDX = UINT32_MAX;
 
 // System representation for a variable-sized overflow value.
 struct overflow_value_t {
